+++ packages/base/bundle.esm.js
@@ -4,17 +4,7 @@ import EventProvider from "./dist/EventProvider.js";
 // ESM bundle targets browsers with native support
 import "./dist/features/OpenUI5Support.js";
 
-// Test components
-import "./test/elements/Accessor.js";
-import "./test/elements/Generic.js";
-import "./test/elements/NoShadowDOM.js";
-import "./test/elements/Parent.js";
-import "./test/elements/Child.js";
-import "./test/elements/WithComplexTemplate.js";
-import "./test/elements/GenericExt.js";
-
 // Test themes - CSS Vars for the sap_fiori_3, sap_fiori_3_dark themes
-import "./test/assets/Themes.js";
 
 // used in test pages
 import { renderFinished } from "./dist/Render.js";
@@ -31,6 +21,8 @@ import { getAnimationMode } from "./dist/config/AnimationMode.js";
 import { getLanguage, setLanguage } from "./dist/config/Language.js";
 import { getCalendarType } from "./dist/config/CalendarType.js";
 import { getTheme, setTheme } from "./dist/config/Theme.js";
+import { getDefaultFontLoading } from "./dist/config/Fonts.js";
+import { getEnableDefaultTooltips } from "./dist/config/Tooltips.js";
 import { getThemeRoot, setThemeRoot } from "./dist/config/ThemeRoot.js";
 import { getNoConflict, setNoConflict } from "./dist/config/NoConflict.js";
 import { getFirstDayOfWeek, getLegacyDateCalendarCustomizing } from "./dist/config/FormatSettings.js";
@@ -53,6 +45,8 @@ window["sap-ui-webcomponents-bundle"] = {
 		getCalendarType,
 		getFirstDayOfWeek,
 		getLegacyDateCalendarCustomizing,
+		getDefaultFontLoading,
+		getEnableDefaultTooltips,
 	},
 	getCurrentRuntimeIndex,
 	getIconNames,