+++ packages/base/test/elements/WithComplexTemplate.d.ts
@@ -1,3 +1,0 @@
-type WithComplexTemplate = any;
-
-export default WithComplexTemplate;