+++ packages/main/src/DayPicker.ts
@@ -172,7 +172,7 @@ class DayPicker extends CalendarPart implements ICalendarPicker {
 	 * When selectionMode="Range" and the first day in the range is selected, this is the currently hovered (when using mouse) or focused (when using keyboard) day by the user
 	 * @private
 	 */
-	 @property()
+	 @property({ type: Number })
 	_secondTimestamp?: number;
 
 	/**
@@ -278,7 +278,6 @@ class DayPicker extends CalendarPart implements ICalendarPicker {
 
 			if (isSelectedBetween) {
 				day.classes += " ui5-dp-item--selected-between";
-
 				day.parts += " day-cell-selected-between";
 			}
 
@@ -290,8 +289,8 @@ class DayPicker extends CalendarPart implements ICalendarPicker {
 				day.classes += " ui5-dp-item--othermonth";
 			}
 
-			if (isWeekend) {
-				day.classes += " ui5-dp-item--weeekend";
+			if ((isWeekend || specialDayType === "NonWorking") && specialDayType !== "Working") {
+				day.classes += " ui5-dp-item--weekend";
 			}
 
 			if (isDisabled) {