+++ packages/main/test/specs/MultiComboBox.spec.js
@@ -541,6 +541,34 @@ describe("MultiComboBox general interaction", () => {
 			assert.strictEqual(tokens.length, 0, "Long token should be deleted" );
 		});
 
+		it("tests if clicking delete icon of a token removes it from the selection (mcb with grouping)", async () => {
+			await browser.url(`test/pages/MultiComboBox.html`);
+
+			const mcb = $("#mcb-grouping");
+			const inner = mcb.shadow$("input");
+
+			await mcb.scrollIntoView();
+			await inner.click();
+
+			await inner.keys("a");
+			const popover = await mcb.shadow$("ui5-responsive-popover");
+			const firstItem = await popover.$(".ui5-multi-combobox-all-items-list > ui5-li");
+			await firstItem.click();
+
+			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
+
+			assert.strictEqual(tokens.length, 1, "Token should be added");
+
+			const token = await mcb.shadow$("ui5-tokenizer ui5-token");
+			const deleteIcon = await token.shadow$(".ui5-token--icon");
+
+			await deleteIcon.click();
+
+			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
+
+			assert.strictEqual(tokens.length, 0, "Token should be deleted");
+		});
+
 		it("prevents selection change event when clicking an item", async () => {
 			await browser.url(`test/pages/MultiComboBox.html`);
 
@@ -1393,6 +1421,26 @@ describe("MultiComboBox general interaction", () => {
 			assert.ok(await tokens[tokens.length - 1].getProperty("focused"), "Last Token is focused");
 		});
 
+		it("Backspace should not delete tokens, when MCB is readonly", async () => {
+			await browser.url(`test/pages/MultiComboBox.html`);
+
+			const mcb = await browser.$("#mcb-ro");
+			const inner = await mcb.shadow$("input");
+			let tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
+
+			await inner.click();
+
+			assert.strictEqual(tokens.length, 3, "3 Tokens are placed in the MCB");
+
+			await inner.keys("Backspace");
+			await inner.keys("Backspace");
+
+			tokens = await mcb.shadow$$(".ui5-multi-combobox-token");
+
+			assert.strictEqual(tokens.length, 3, "3 Tokens are placed in the MCB");
+			assert.ok(await tokens[tokens.length - 1].getProperty("focused"), "Last Token is focused");
+		});
+
 		it("should open/close popover on keyboard combination ctrl + i", async () => {
 			const mcb = await browser.$("#truncated-token");
 			const inner = await mcb.shadow$("input");
@@ -1499,6 +1547,35 @@ describe("MultiComboBox general interaction", () => {
 
 			await mcb.click();
 			assert.notOk(await popover.getProperty("open"), "Popover with valueStateMessage should not be opened.");
+		});
+
+		it("Should not open value state popup when popover is open after clicking n more link", async () => {
+			const mcb = await browser.$("#mcb-select-all-vs");
+
+			await mcb.scrollIntoView();
+			await mcb.click();
+
+			const popover = await mcb.shadow$("ui5-responsive-popover");
+			const valueStatePopover = await mcb.shadow$("ui5-popover");
+
+			assert.notOk(await popover.getProperty("open"), "Popover with value state message header should not be opened.");
+			assert.ok(await valueStatePopover.getProperty("open"), "Value state popover should be opened.");
+
+			await mcb.keys("F4");
+			await browser.keys("ArrowUp");
+			await browser.keys("Enter");
+			await browser.keys("Tab");
+
+			const nMoreText = await mcb.shadow$("ui5-tokenizer").shadow$(".ui5-tokenizer-more-text");
+			nMoreText.click();
+
+			await browser.waitUntil(async () => mcb.getProperty("open"), {
+				timeout: 1000,
+				timeoutMsg: "Popover is open"
+			});
+
+			assert.ok(await popover.getProperty("open"), "Popover with value state message header should be opened.");
+			assert.notOk(await valueStatePopover.getProperty("open"), "Value state popover should not be opened.");
 		});
 
 		it("Should apply correct text to the tokens overflow indicator", async () => {
@@ -1699,6 +1776,16 @@ describe("MultiComboBox general interaction", () => {
 			await mcb.scrollIntoView();
 
 			assert.strictEqual(await innerInput.getAttribute("aria-label"), await mcbLabel.getHTML(false), "aria-label attribute is correct.");
+		});
+
+		it("Should apply aria-controls pointing to the responsive popover", async () => {
+			const mcb = await browser.$("#mcb-predefined-value");
+			const innerInput = await mcb.shadow$("input");
+			const popover = await mcb.shadow$("ui5-responsive-popover");
+
+			await mcb.scrollIntoView();
+
+			assert.strictEqual(await innerInput.getAttribute("aria-controls"), await popover.getAttribute("id"), "aria-controls attribute is correct.");
 		});
 
 		it("Should render aria-haspopup attribute with value 'dialog'", async () => {